gcp/pyproject.toml
[project]
name = "gcp"
version = "0.1.0"
description = ""
authors = [
    {name = "uuboyscy",email = "aegis12321@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-bigquery (>=3.31.0,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "db-dtypes (>=1.4.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
